Update null tests
