--- modules/ballerina-native/src/test/resources/lang/values/null/arithmatic-operation-on-null.bal
@@ -1,3 +1,0 @@
-function testLogicalOperationOnNull() (any) {
-    return (null + null);
-}
