--- modules/ballerina-native/src/test/resources/lang/values/null/invalid-function-call-with-nulll.bal
@@ -1,8 +1,0 @@
-function testInvalidFunctionCallWithNull() (any) {
-    return foo(null);
-}
-
-function foo(string s) {
-    return s;
-}
-
