--- modules/ballerina-native/src/test/resources/lang/values/null/logical-operation-on-null.bal
@@ -1,4 +1,0 @@
-function testLogicalOperationOnNull() (boolean) {
-    xml x;
-    return (null > x);
-}
