--- modules/ballerina-native/src/test/resources/lang/values/null/null-cast-to-reference-type.bal
@@ -1,3 +1,0 @@
-function testNullForValueType() {
-    json j = (json) null;
-}
