--- modules/ballerina-native/src/test/resources/lang/values/null/null-cast-to-value-type.bal
@@ -1,3 +1,0 @@
-function testNullForValueType() {
-    string s = (string) null;
-}
