--- modules/ballerina-native/src/test/resources/lang/values/null/null-for-value-types.bal
@@ -1,3 +1,0 @@
-function testNullForValueType() {
-    int a = null;
-}
