--- modules/ballerina-native/src/test/resources/lang/values/null/null-type-var.bal
@@ -1,3 +1,0 @@
-function testCreateNullTypeVar() {
-    null x;
-}
