+++ modules/ballerina-test/src/test/resources/test-src/types/null/arithmatic-operation-on-null.bal
@@ -1,0 +1,3 @@
+function testLogicalOperationOnNull () (any) {
+    return (null + null);
+}
