+++ modules/ballerina-test/src/test/resources/test-src/types/null/invalid-function-call-with-nulll.bal
@@ -1,0 +1,7 @@
+function testInvalidFunctionCallWithNull () (any) {
+    return foo(null);
+}
+
+function foo (string s) {
+    return s;
+}
