+++ modules/ballerina-test/src/test/resources/test-src/types/null/logical-operation-on-null.bal
@@ -1,0 +1,4 @@
+function testLogicalOperationOnNull () (boolean) {
+    xml x;
+    return (null > x);
+}
