+++ modules/ballerina-test/src/test/resources/test-src/types/null/null-cast-to-reference-type.bal
@@ -1,0 +1,3 @@
+function testNullForValueType () {
+    json j = (json)null;
+}
