+++ modules/ballerina-test/src/test/resources/test-src/types/null/null-cast-to-value-type.bal
@@ -1,0 +1,3 @@
+function testNullForValueType () {
+    string s = (string)null;
+}
