+++ modules/ballerina-test/src/test/resources/test-src/types/null/null-for-value-types.bal
@@ -1,0 +1,3 @@
+function testNullForValueType () {
+    int a = null;
+}
