+++ modules/ballerina-test/src/test/resources/test-src/types/null/null-of-different-types.bal
@@ -1,10 +1,10 @@
-function testCompareNullOfDifferentTypes()(boolean, xml, json) {
+function testCompareNullOfDifferentTypes () (boolean, xml, json) {
     xml x = null;
     json j = null;
     int a = 0;
     if (x == j) {
         a = 11;
     }
-    
+
     return x, j, a;
 }
