+++ modules/ballerina-test/src/test/resources/test-src/types/null/null-type-var.bal
@@ -1,0 +1,3 @@
+function testCreateNullTypeVar () {
+    null x;
+}
